fix: harden NIP-46 bunker security and correctness
- Validate bunker connect secret before granting access
- Default to rejecting requests when no approval callback
- Restrict default permissions to GET_PUBLIC_KEY only
- Parse and apply requested permissions from connect params
- URL-encode bunker secret to prevent URL parsing issues
- Cap connected apps at 100 and logs at 1000 entries
- Wrap blocking recv_timeout in block_in_place
- Fix UTF-8 panic on multi-byte content truncation
- Widen event Kind to u32 for full NIP compatibility
- Return error on invalid tags instead of silently dropping

+++ keep-nip46/src/types.rs
@@ -60,7 +60,7 @@ impl Nip46Response {
 
 #[derive(Debug, serde::Deserialize)]
 pub(crate) struct PartialEvent {
-    pub kind: u16,
+    pub kind: u32,
     pub content: String,
     #[serde(default)]
     pub tags: Vec<Vec<String>>,
